userNo 정의를 long에서 int로 변경

+++ src/main/java/org/choongang/students/mapper/StuinfoMapper.java
@@ -6,10 +6,10 @@ import java.util.List;
 
 public interface StuinfoMapper {
     List<StuInfo> getList();
-    List<StuInfo> get(long userNo);
+    List<StuInfo> get(int userNo);
     int register(StuInfo stuinfo);
-    int delete(long userNo);
+    int delete(int userNo);
     int modify(StuInfo stuinfo);
-    int exists(long userNo);
+    int exists(int userNo);
 }
